# ENTREGABLES_SIGAI_SES/01_DOCUMENTACION_TECNICA/API_SPEC/openapi.yaml
openapi: 3.0.3
info:
  title: SIGAI-SES API
  version: "1.0.0"
  description: >
    API RESTful del Sistema Integral de Gestión de Activos e Inventario
    para Seguridad Electrónica Securitas Colombia.
  contact:
    name: Equipo de Desarrollo SIGAI-SES
    email: soporte@ses-securitas.com.co
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.ses.example.com/v1
    description: Producción
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido de /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id_usuario:
          type: integer
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          type: string
          enum: [ADMIN, TECNICO, TECNICO_LABORATORIO]
        id_regional:
          type: integer
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [nombre, email, password, rol]
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        rol:
          type: string
          enum: [ADMIN, TECNICO, TECNICO_LABORATORIO]
        id_regional:
          type: integer
          nullable: true
        cedula:
          type: string
        codigo_empleado:
          type: string

    # ==================== ITEMS ====================
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
        prioridad:
          type: string
          enum: [critica, alta, media, baja]
        estado:
          type: string
          enum: [activa, reconocida, resuelta, ignorada]
        titulo:
          type: string
        descripcion:
          type: string
        item_nombre:
          type: string
        valor_actual:
          type: number
        valor_umbral:
          type: number
        created_at:
          type: string
          format: date-time

    # ==================== AUDITORIA ====================
    AuditLog:
      type: object
      properties:
        id_log:
          type: integer
        id_usuario:
          type: integer
        tabla_afectada:
          type: string
        accion:
          type: string
          enum: [CREATE, UPDATE, DELETE, LOGIN]
        id_registro:
          type: integer
        valor_anterior:
          type: string
        valor_nuevo:
          type: string
        fecha_accion:
          type: string
          format: date-time

    # ==================== ERRORES ====================
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      summary: Listar usuarios (ADMIN)
      tags: [Users]
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Crear usuario (ADMIN)
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id_user}:
    put:
      summary: Actualizar usuario
      tags: [Users]
      parameters:
        - name: id_user
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: Usuario actualizado
    delete:
      summary: Eliminar usuario (ADMIN)
      tags: [Users]
      parameters:
        - name: id_user
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Usuario eliminado
  /users/audit:
    get:
      summary: Ver logs de auditoria
      tags: [Users]
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Logs de auditoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  # ==================== INVENTORY ====================
...
  /alerts:
    get:
      summary: Listar alertas
      tags: [Alerts]
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [activa, reconocida, resuelta, ignorada]
        - name: prioridad
          in: query
          schema:
            type: string
            enum: [critica, alta, media, baja]
        - name: tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      summary: Crear alerta manual
      tags: [Alerts]
      responses:
        "201":
          description: Alerta creada
...